Moved some files to separate folders.
