+++ floppies/l-system-generator/source/library/includes/system/rule.h
