+++ floppies/l-system-generator/source/library/includes/system/symbol.h
