+++ floppies/l-system-generator/source/library/system/symbol.c
@@ -1,4 +1,4 @@
-#include "rule.h"
+#include "system/rule.h"
 #include "memory/safe_alloc.h"
 
 SymbolList alloc_empty_symbol_list()
